requirements.txt
absl-py==0.9.0
appdirs==1.4.3
astunparse==1.6.3
attrs==19.3.0
black==19.10b0
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
future==0.18.2
gast==0.3.3
google-auth==1.14.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
Keras-Preprocessing==1.1.0
Markdown==3.2.1
mypy==0.770
mypy-extensions==0.4.3
numpy==1.18.4
oauthlib==3.1.0
opt-einsum==3.2.1
pandas==1.0.3
pathspec==0.8.0
Pillow==7.1.2
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2020.1
regex==2020.4.4
requests==2.23.0
requests-oauthlib==1.3.0
rope==0.16.0
rsa==4.0
scipy==1.4.1
six==1.14.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
tensorflow==2.2.0rc4
tensorflow-estimator==2.2.0
termcolor==1.1.0
toml==0.10.0
torch==1.5.0
torchvision==0.6.0
tqdm==4.46.0
typed-ast==1.4.1
typing-extensions==3.7.4.2
ulid-py==0.0.14
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
